Функции получения response перенесены в модуль httpclient.py
